backend/requirements.txt
django>=4.2.0,<4.3.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
djangorestframework
pyyaml
requests
django-cors-headers